+++ Backend/HireAProBackend/Services/ISaveToken.cs
@@ -1,0 +1,9 @@
+﻿using Microsoft.AspNetCore.Mvc;
+
+namespace HireAProBackend.Services
+{
+    public interface ISaveToken
+    {
+        Task<bool> GuardarTokenAsync(string tokenEntrada, string correo);
+    }
+}
